Added single cam draw ui element

+++ Tankerfield/Tankerfield/M_Map.cpp
@@ -675,8 +675,8 @@ fPoint M_Map::MapToCamera(const fPoint map_pos, const Camera* camera)
 {
 	fPoint camera_pos(0.0F, 0.0F);
 	
-	camera_pos.x = (map_pos.x - map_pos.y) * (data.tile_width * 0.5f)  - camera->rect.x + camera->viewport.x ;
-	camera_pos.y = (map_pos.x + map_pos.y) * (data.tile_height * 0.5f) - camera->rect.y + camera->viewport.y;
+	camera_pos.x = (map_pos.x - map_pos.y) * ((float)data.tile_width * 0.5f)  - (float)camera->rect.x + (float)camera->viewport.x ;
+	camera_pos.y = (map_pos.x + map_pos.y) * ((float)data.tile_height * 0.5f) - (float)camera->rect.y + (float)camera->viewport.y;
 
 	return camera_pos;
 }
